refactor: update function for lineItem - add possibility to remove item
